requirements.txt
streamlit
sentence-transformers
scikit-learn
PyMuPDF
llama-cpp-python
numpy
transformers
torch
requests